requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
beautifulsoup4
selenium 
webdriver-manager 
selenium-stealth
python-dotenv
openai